requirements.txt
certifi==2017.11.5
chardet==3.0.4
Django==1.11.3
django-oauth2-provider==0.2.6.1
djangorestframework==3.6.4
Faker==0.8.4
idna==2.6
Markdown==2.6.9
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
shortuuid==0.5.0
six==1.11.0
urllib3==1.22
